initializeInfoWindow,openInfoWindow,setContentOfInfoWindow,initializePolyline,updatePolyline,clickFunction functions are added in GMapOperator.js
initializeInfoWindow initializes the window by the string that is given.And returns the infoWindow.
openInfoWindow opens the infoWindow that is given.
setContentOfInfoWindow updates the string of infoWindow that is given.

initializePolyline creates and initializes the PolyLine.
updatePolyline updates the polyLine by new Location that are given.

clickFunction calls the function that is given when map is clicked.