According to "Comment on revision r561" Bug 2 is fixed.
Bug-2: if i fill everything and click register, it registers and writes  "{"result":"1"}{"result":"1"}" on forum.

For Bug 3, device id is used as email field.

Bug-3: if i enter something a name and device id, it is ok, it registers correctly. But if i use same name in registering another device, it gives error in background. Problem is due to email field duplicate data.